chore: Corrected handling of background images
