+++ src/Uno.UI/UI/Xaml/Controls/ContentPresenter/ContentPresenter.cs
@@ -42,6 +42,13 @@ using ViewGroup = Microsoft.UI.Xaml.UIElement;
 
 namespace Microsoft.UI.Xaml.Controls
 {
+	/// <summary>
+	/// Declares a Content presenter
+	/// </summary>
+	/// <remarks>
+	/// The content presenter is used for compatibility with WPF concepts,
+	/// but the ContentSource property is not available, because there are ControlTemplates for now.
+	/// </remarks>
 	[ContentProperty(Name = "Content")]
 	public partial class ContentPresenter : FrameworkElement, IFrameworkTemplatePoolAware
 #if !__CROSSRUNTIME__ && !IS_UNIT_TESTS
